UserMenu component (4651)
