+++ jmix-security/security-flowui/src/main/java/io/jmix/securityflowui/SecurityUiProperties.java
@@ -1,0 +1,42 @@
+/*
+ * Copyright 2020 Haulmont.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.jmix.securityflowui;
+
+import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.boot.context.properties.bind.DefaultValue;
+
+@ConfigurationProperties(prefix = "jmix.security.ui")
+public class SecurityUiProperties {
+
+    /**
+     * Defines the maximum number of users that can be substituted by the currently authenticated user,
+     * which shows a submenu instead of opening a lookup view.
+     */
+    int substituteUserActionMaxSubstitutions;
+
+    public SecurityUiProperties(
+            @DefaultValue("5") int substituteUserActionMaxSubstitutions) {
+        this.substituteUserActionMaxSubstitutions = substituteUserActionMaxSubstitutions;
+    }
+
+    /**
+     * @see #substituteUserActionMaxSubstitutions
+     */
+    public int getSubstituteUserActionMaxSubstitutions() {
+        return substituteUserActionMaxSubstitutions;
+    }
+}
